Activities

+++ app/src/main/java/com/elliemoritz/coinbook/domain/entities/Category.kt
@@ -4,5 +4,5 @@ import com.elliemoritz.coinbook.domain.entities.helpers.UNDEFINED_ID
 
 data class Category(
     val name: String,
-    val id: Int = UNDEFINED_ID
+    val id: Long = UNDEFINED_ID
 )
