+++ app/src/main/java/com/elliemoritz/coinbook/domain/repositories/LimitsRepository.kt
@@ -4,11 +4,14 @@ import com.elliemoritz.coinbook.domain.entities.Limit
 import kotlinx.coroutines.flow.Flow
 
 interface LimitsRepository {
+
     fun getLimitsList(): Flow<List<Limit>>
-    fun getLimit(id: Int): Flow<Limit>
-    fun getLimitByCategoryId(categoryId: Int): Flow<Limit?>
+
+    fun getLimit(id: Long): Flow<Limit>
+    fun getLimitByCategoryId(categoryId: Long): Flow<Limit?>
     suspend fun addLimit(limit: Limit)
     suspend fun editLimit(limit: Limit)
     suspend fun removeLimit(limit: Limit)
-    fun getLimitsCount(): Flow<Int>
+
+    fun getLimitsCount(): Flow<Long>
 }
